+++ app/models/induction_programme.rb
@@ -25,8 +25,9 @@ class InductionProgramme < ApplicationRecord
   has_one :lead_provider, through: :partnership
   has_one :delivery_partner, through: :partnership
   has_one :cpd_lead_provider, through: :lead_provider
+  has_one :school, through: :school_cohort
+  has_one :cohort, through: :school_cohort
 
-  delegate :cohort, :cohort_id, :school, to: :school_cohort
   delegate :name, to: :core_induction_programme, allow_nil: true, prefix: true
 
   after_commit :touch_induction_records
@@ -40,7 +41,7 @@ class InductionProgramme < ApplicationRecord
   end
 
   def same_induction_as?(other)
-    school_cohort.cohort == other.school_cohort.cohort &&
+    cohort == other.cohort &&
       training_programme == other.training_programme &&
       lead_provider == other.lead_provider &&
       delivery_partner == other.delivery_partner &&
