+++ app/services/coc_set_participant_categories.rb
@@ -1,142 +1,222 @@
 # frozen_string_literal: true
 
 class CocSetParticipantCategories < BaseService
-  ParticipantCategories = Struct.new(:eligible, :ineligible, :withdrawn, :contacted_for_info, :details_being_checked, :transferring_in, :transferring_out, :transferred, :no_qts_participants)
+  ParticipantCategories = Struct.new(
+    :eligible,
+    :contacted_for_info,
+    :details_being_checked,
+    :ineligible,
+    :no_qts_participants,
+    :transferred,
+    :transferring_in,
+    :transferring_out,
+    :withdrawn,
+  )
 
-  # replace [] in call back to transferring_out_participants method once
-  # transferring out journey is done
-  # transferring out ppts are being shown in active_induction_records for now
-  def call
-    ParticipantCategories.new(
-      eligible_participants,
-      ineligible_participants,
-      withdrawn_participants,
-      contacted_for_info_participants,
-      details_being_checked,
-      transferring_in_participants,
-      transferring_out_participants,
-      transferred_participants,
-      no_qts_participants,
-    )
+  def ect_categories
+    @ect_categories ||= categories(ParticipantProfile::ECT.name)
+  end
+
+  def ineligible
+    @ineligible ||= ect_categories.ineligible + mentor_categories.ineligible
+  end
+
+  def mentor_categories
+    @mentor_categories ||= categories(ParticipantProfile::Mentor.name)
+  end
+
+  def transferred
+    @transferred ||= ect_categories.transferred + mentor_categories.transferred
+  end
+
+  def transferring_in
+    @transferring_in ||= ect_categories.transferring_in + mentor_categories.transferring_in
+  end
+
+  def transferring_out
+    @transferring_out ||= ect_categories.transferring_out + mentor_categories.transferring_out
+  end
+
+  def withdrawn
+    @withdrawn ||= ect_categories.withdrawn + mentor_categories.withdrawn
   end
 
 private
 
-  attr_reader :school_cohort, :user, :profile_type
+  attr_reader :school_cohort, :user
 
-  def initialize(school_cohort, user, profile_type)
+  def initialize(school_cohort, user)
     @school_cohort = school_cohort
     @user = user
-    @profile_type = profile_type
   end
 
-  def ineligible_participants
-    active_induction_records
-      .fip
-      .merge(profile_type.ineligible_status)
-  end
-
-  def eligible_participants
-    fip_eligible_participants + cip_eligible_participants
-  end
-
-  def withdrawn_participants
-    withdrawn_induction_records
-  end
-
-  def contacted_for_info_participants
-    active_induction_records
-      .merge(profile_type.contacted_for_info)
-  end
-
-  def details_being_checked_participants
-    active_induction_records
-      .fip
-      .merge(profile_type.details_being_checked)
-  end
-
-  def details_being_checked
-    details_being_checked_participants - no_qts_participants
-  end
-
-  def no_qts_participants
-    details_being_checked_participants.select { |record| record.participant_profile.ecf_participant_eligibility&.no_qts_reason? }
-  end
-
-  def transferring_in_participants
-    InductionRecordPolicy::Scope.new(
-      user,
-      school_cohort
-        .transferring_in_induction_records
-        .joins(:participant_profile)
-        .where(participant_profiles: { type: profile_type.to_s })
-        .includes(:user)
-        .order("users.full_name"),
-    ).resolve
-  end
-
-  def transferring_out_participants
-    InductionRecordPolicy::Scope.new(
-      user,
-      school_cohort
-        .transferring_out_induction_records
-        .joins(:participant_profile)
-        .where(participant_profiles: { type: profile_type.to_s })
-        .includes(:user)
-        .order("users.full_name"),
-    ).resolve
-  end
-
-  def transferred_participants
-    InductionRecordPolicy::Scope.new(
-      user,
-      school_cohort
-        .transferred_induction_records
-        .joins(:participant_profile)
-        .where(participant_profiles: { type: profile_type.to_s })
-        .includes(:user)
-        .order("users.full_name"),
-    ).resolve
-  end
-
+  # Query methods
   def active_induction_records
-    InductionRecordPolicy::Scope.new(
+    @active_induction_records ||= induction_records
+                                    .select { |induction_record| active_induction_record?(induction_record) }
+  end
+
+  def categories(profile_type)
+    ParticipantCategories.new(
+      eligible_participants(profile_type),
+      contacted_for_info_participants(profile_type),
+      details_being_checked(profile_type),
+      ineligible_participants(profile_type),
+      no_qts_participants(profile_type),
+      transferred_participants(profile_type),
+      transferring_in_participants(profile_type),
+      transferring_out_participants(profile_type),
+      withdrawn_participants(profile_type),
+    )
+  end
+
+  def cip_eligible_participants(profile_type)
+    @cip_eligible_participants ||= active_induction_records
+                                     .select { |induction_record| cip_eligible_participant?(induction_record) }
+                                     .group_by(&:participant_type)
+
+    Array(@cip_eligible_participants[profile_type])
+  end
+
+  def contacted_for_info_participants(profile_type)
+    @contacted_for_info_participants ||= active_induction_records
+                                           .select { |induction_record| contacted_for_info_participant?(induction_record) }
+                                           .group_by(&:participant_type)
+
+    Array(@contacted_for_info_participants[profile_type])
+  end
+
+  # Retrieve the relevant induction record for each participant in the school cohort:
+  # For participants transferring into the school cohort, get their transferring_in? induction_record.
+  # For participants transferred from another school, get their transferred? induction_record
+  # For the rest of participants, get their current induction record
+  def induction_records
+    @induction_records ||= InductionRecordPolicy::Scope.new(
       user,
       school_cohort
-        .active_induction_records
-        .or(InductionRecord.claimed_by_another_school)
-        .where.not(training_status: :withdrawn)
-        .joins(:participant_profile)
-        .where(participant_profiles: { type: profile_type.to_s })
-        .includes(:user)
+        .induction_records
+        .current_or_transferring_in_or_transferred
+        .eager_load(induction_programme: %i[school core_induction_programme lead_provider delivery_partner],
+                    participant_profile: %i[user ecf_participant_eligibility ecf_participant_validation_data])
         .order("users.full_name"),
-    ).resolve
+    ).resolve.to_a
   end
 
-  def withdrawn_induction_records
-    InductionRecordPolicy::Scope.new(
-      user,
-      school_cohort
-        .current_induction_records
-        .training_status_withdrawn
-        .joins(:participant_profile)
-        .where(participant_profiles: { type: profile_type.to_s })
-        .includes(:user)
-        .order("users.full_name"),
-    ).resolve
+  def details_being_checked(profile_type)
+    details_being_checked_participants(profile_type) - no_qts_participants(profile_type)
   end
 
-  def fip_eligible_participants
-    active_induction_records
-      .fip
-      .merge(profile_type.eligible_status)
+  def details_being_checked_participants(profile_type = nil)
+    @details_being_checked_participants ||= active_induction_records
+                                              .select { |induction_record| details_being_checked_participant?(induction_record) }
+                                              .group_by(&:participant_type)
+
+    if profile_type
+      Array(@details_being_checked_participants[profile_type])
+    else
+      @details_being_checked_participants.values.flatten
+    end
   end
 
-  def cip_eligible_participants
-    # find all the participants that have attempted to validate in any core_induction_programme
-    # for the school_cohort
-    active_induction_records
-      .joins(participant_profile: :ecf_participant_validation_data)
-      .cip
+  def eligible_participants(profile_type)
+    (fip_eligible_participants(profile_type) + cip_eligible_participants(profile_type))
+      .sort_by(&:participant_full_name)
+  end
+
+  def fip_eligible_participants(profile_type)
+    @fip_eligible_participants ||= active_induction_records
+                                     .select { |induction_record| fip_eligible_participant?(induction_record) }
+                                     .group_by(&:participant_type)
+
+    Array(@fip_eligible_participants[profile_type])
+  end
+
+  def ineligible_participants(profile_type)
+    @ineligible_participants ||= active_induction_records
+                                   .select { |induction_record| ineligible_participant?(induction_record) }
+                                   .group_by(&:participant_type)
+
+    Array(@ineligible_participants[profile_type])
+  end
+
+  def no_qts_participants(profile_type)
+    @no_qts_participants ||= details_being_checked_participants
+                               .select { |induction_record| no_qts_participant?(induction_record) }
+                               .group_by(&:participant_type)
+
+    Array(@no_qts_participants[profile_type])
+  end
+
+  def transferred_participants(profile_type)
+    @transferred_participants ||= induction_records
+                                    .select(&:transferred?)
+                                    .group_by(&:participant_type)
+
+    Array(@transferred_participants[profile_type])
+  end
+
+  def transferring_in_participants(profile_type)
+    @transferring_in_participants ||= induction_records
+                                        .select(&:transferring_in?)
+                                        .group_by(&:participant_type)
+
+    Array(@transferring_in_participants[profile_type])
+  end
+
+  def transferring_out_participants(profile_type)
+    @transferring_out_participants ||= induction_records
+                                         .select(&:transferring_out?)
+                                         .group_by(&:participant_type)
+
+    Array(@transferring_out_participants[profile_type])
+  end
+
+  def withdrawn_participants(profile_type)
+    @withdrawn_participants ||= induction_records
+                                  .select { |induction_record| withdrawn_participant?(induction_record) }
+                                  .group_by(&:participant_type)
+
+    Array(@withdrawn_participants[profile_type])
+  end
+
+  # Type check methods
+  def active_induction_record?(induction_record)
+    !induction_record.training_status_withdrawn? &&
+      (induction_record.active? || induction_record.claimed_by_another_school?)
+  end
+
+  def cip_eligible_participant?(induction_record)
+    induction_record.enrolled_in_cip? &&
+      induction_record.participant_profile.completed_validation_wizard?
+  end
+
+  def contacted_for_info_participant?(induction_record)
+    induction_record.participant_profile.contacted_for_info?
+  end
+
+  def details_being_checked_participant?(induction_record)
+    induction_record.enrolled_in_fip? &&
+      induction_record.participant_profile.manual_check_needed?
+  end
+
+  def fip_eligible_participant?(induction_record)
+    induction_record.enrolled_in_fip? &&
+      (induction_record.participant_profile.fundable? ||
+        induction_record.participant_profile.ineligible_and_duplicated_or_previously_participated?)
+  end
+
+  def ineligible_participant?(induction_record)
+    induction_record.enrolled_in_fip? &&
+      induction_record.participant_profile.ineligible_but_not_duplicated_or_previously_participated?
+  end
+
+  def no_qts_participant?(induction_record)
+    induction_record.participant_profile.ecf_participant_eligibility&.no_qts_reason?
+  end
+
+  def withdrawn_participant?(induction_record)
+    induction_record.training_status_withdrawn? &&
+      !induction_record.transferring_in? &&
+      !induction_record.transferred?
   end
 end
